Extract service client functions for re-use
We need to start breaking up the client into more manageable parts, each
related to the service it interacts with.
