+++ src/azure/resource-manager/ResourceManagerClient.ts
@@ -1,4 +1,3 @@
-/* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable @typescript-eslint/no-non-null-assertion */
 
 import { ComputeManagementClient } from "@azure/arm-compute";
@@ -37,133 +36,11 @@ import {
 } from "@azure/arm-sql/esm/models";
 import { StorageManagementClient } from "@azure/arm-storage";
 import { BlobContainer, StorageAccount } from "@azure/arm-storage/esm/models";
-import {
-  exponentialRetryPolicy,
-  HttpResponse,
-  RequestPolicyFactory,
-  systemErrorRetryPolicy,
-  throttlingRetryPolicy,
-} from "@azure/ms-rest-js";
-import {
-  IntegrationLogger,
-  IntegrationError,
-} from "@jupiterone/jupiter-managed-integration-sdk";
-import { retry } from "@lifeomic/attempt";
 
-import { AzureIntegrationInstanceConfig } from "../../types";
 import { resourceGroupName } from "../utils";
-import authenticate from "./authenticate";
-import { bunyanLogPolicy } from "./BunyanLogPolicy";
-import { AzureManagementClientCredentials } from "./types";
+import { Client, iterateAllResources, ListResourcesEndpoint } from "./client";
 
-/**
- * An Azure resource manager endpoint that has `listAll` and `listAllNext` functions.
- */
-interface ListAllResourcesEndpoint {
-  listAll: <ListResponseType>() => Promise<ListResponseType>;
-  listAllNext: <ListResponseType>(
-    nextLink: string,
-  ) => Promise<ListResponseType>;
-}
-
-/**
- * An Azure resource manager endpoint that has `list` and `listNext` functions.
- */
-interface ListResourcesEndpoint {
-  list<ListResponseType>(): Promise<ListResponseType>;
-  listNext<ListResponseType>(nextLink: string): Promise<ListResponseType>;
-}
-
-interface ResourceListResponse<T> extends Array<T> {
-  readonly _response: HttpResponse;
-  readonly nextLink?: string;
-}
-
-/**
- * Retries a resource request that throws an error on a 429 response.
- *
- * Note that the documentation at
- * https://docs.microsoft.com/en-us/azure/virtual-machines/troubleshooting/troubleshooting-throttling-errors
- * states plainly:
- *
- * > In high-volume API automation cases, consider implementing proactive
- * > client-side self-throttling when the available call count for a target
- * > operation group drops below some low threshold.
- *
- * That would be nice for them to handle in the client, and it would be great to
- * see the endpoints always return the resource provider counts remaining, and
- * they'd get bonus points for having useful `retry-after` header values.
- *
- * Some resource provider endpoints have returned a 429 response with a
- * `retry-after` that is useless, and another try gets another 429, which the
- * `ThrottlingRetryPolicy` will not handle. That is, given 100 req/5 minutes, a
- * burst of 100 requests will burn the allowed requests, a 429 is returned with
- * `retry-after: 17`, but in fact the API will not answer again until up to 5
- * minutes has passed!
- *
- * The approach here is designed to:
- *
- * 1. Allow a burst of requests up to the maximum per period supported by the
- *    resource provider vs. always spreading out requests, which makes scenarios
- *    where the endpoint is only called 10 times really slow.
- * 2. Wait the period out after two subsequent 429 responses, because once we've
- *    seen a 429, it means we've burned up our limit for the period.
- * 3. Allow resource providers that return a useful `retry-after` to continue to
- *    work, because the `ThrottlingRetryPolicy` is still in place.
- * 4. Allow specific resource provider endpoint period wait times.
- * 5. Allow other error responses to continue to throw an error (this will not
- *    attempt to retry those).
- *
- * @param requestFunc code making a request to Azure RM
- * @param endpointRatePeriod the resource provider's rate limiting period; the
- * time in a reqs/time ratio
- */
-/* istanbul ignore next: testing iteration might be difficult */
-function retryResourceRequest<ResponseType>(
-  requestFunc: () => Promise<ResponseType>,
-  endpointRatePeriod: number,
-): Promise<ResponseType> {
-  return retry(
-    async _context => {
-      return requestFunc();
-    },
-    {
-      // Things aren't working as expected if we're asked to retry more than
-      // once. The request policy of the `ServiceClient` will handle all other
-      // retry scenarios.
-      maxAttempts: 2,
-
-      // No delay on the first attempt, wait for next period on subsequent
-      // attempts. Assumes non-429 responses will not lead to subsequent
-      // attempts (`handleError` will abort for other error responses).
-      calculateDelay: (context, _options) => {
-        return context.attemptNum === 0 ? 0 : endpointRatePeriod;
-      },
-
-      // Most errors will be handled by the request policies. They will raise
-      // a `RestError.statusCode: 429` when they see two 429 responses in a row,
-      // which is the scenario we're aiming to address with our retry.
-      handleError: (err, context, _options) => {
-        if (err.statusCode !== 429) {
-          context.abort();
-        }
-      },
-    },
-  );
-}
-
-export default class ResourceManagerClient {
-  private auth: AzureManagementClientCredentials;
-  private serviceClientCache: Map<any, any>;
-
-  constructor(
-    private config: AzureIntegrationInstanceConfig,
-    readonly logger: IntegrationLogger,
-    readonly noRetryPolicy: boolean = false,
-  ) {
-    this.serviceClientCache = new Map();
-  }
-
+export default class ResourceManagerClient extends Client {
   //// Compute and Network ////
 
   public async iterateNetworkInterfaces(
@@ -172,7 +49,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       NetworkManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.networkInterfaces,
       callback,
@@ -185,7 +63,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       ComputeManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.virtualMachines,
       callback,
@@ -199,7 +78,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       ComputeManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.images,
       callback,
@@ -225,7 +105,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       NetworkManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.virtualNetworks,
       callback,
@@ -239,7 +120,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       NetworkManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.loadBalancers,
       callback,
@@ -252,7 +134,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       NetworkManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.networkSecurityGroups,
       callback,
@@ -265,7 +148,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       NetworkManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.publicIPAddresses,
       callback,
@@ -280,7 +164,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       StorageManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.storageAccounts,
       callback,
@@ -297,7 +182,8 @@ export default class ResourceManagerClient {
     const resourceGroup = resourceGroupName(storageAccount.id, true)!;
     const accountName = storageAccount.name!;
 
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: {
         list: async () => {
@@ -308,7 +194,7 @@ export default class ResourceManagerClient {
         ) => {
           return serviceClient.blobContainers.listNext(nextLink);
         },
-      } as any,
+      } as ListResourcesEndpoint,
       callback,
     });
   }
@@ -321,7 +207,8 @@ export default class ResourceManagerClient {
     const serviceClient = await this.getAuthenticatedServiceClient(
       SqlManagementClient,
     );
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: serviceClient.servers,
       callback,
@@ -341,7 +228,8 @@ export default class ResourceManagerClient {
     const resourceGroup = resourceGroupName(server.id, true)!;
     const serverName = server.name!;
 
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: {
         list: async () => {
@@ -350,7 +238,7 @@ export default class ResourceManagerClient {
             serverName,
           );
         },
-      } as any,
+      } as ListResourcesEndpoint,
       callback,
     });
   }
@@ -383,7 +271,8 @@ export default class ResourceManagerClient {
     const resourceGroup = resourceGroupName(server.id, true)!;
     const serverName = server.name!;
 
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: {
         list: async () => {
@@ -392,7 +281,7 @@ export default class ResourceManagerClient {
             serverName,
           );
         },
-      } as any,
+      } as ListResourcesEndpoint,
       callback,
     });
   }
@@ -427,7 +316,8 @@ export default class ResourceManagerClient {
     const resourceGroup = resourceGroupName(server.id, true)!;
     const serverName = server.name!;
 
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: {
         list: async () => {
@@ -436,7 +326,7 @@ export default class ResourceManagerClient {
             serverName,
           );
         },
-      } as any,
+      } as ListResourcesEndpoint,
       callback,
     });
   }
@@ -471,7 +361,8 @@ export default class ResourceManagerClient {
     const resourceGroup = resourceGroupName(server.id, true)!;
     const serverName = server.name!;
 
-    return this.iterateAllResources({
+    return iterateAllResources({
+      logger: this.logger,
       serviceClient,
       resourceEndpoint: {
         list: async () => {
@@ -480,134 +371,8 @@ export default class ResourceManagerClient {
             serverName,
           );
         },
-      } as any,
+      } as ListResourcesEndpoint,
       callback,
     });
-  }
-
-  //// Private Functions ////
-
-  private async getAuthenticatedServiceClient<T>(ctor: {
-    new (...args: any[]): T;
-  }): Promise<T> {
-    let client = this.serviceClientCache.get(ctor);
-    if (!client) {
-      client = await this.createAuthenticatedClient(ctor, this.config);
-      this.serviceClientCache.set(ctor, client);
-    }
-    return client as T;
-  }
-
-  private async createAuthenticatedClient<T>(
-    ctor: { new (...args: any[]): T },
-    config: AzureIntegrationInstanceConfig,
-  ): Promise<T> {
-    if (!this.auth) {
-      this.auth = await authenticate(config);
-    }
-
-    // Builds a custom policy chain to address
-    // https://github.com/Azure/azure-sdk-for-js/issues/7989
-    // See also: https://docs.microsoft.com/en-us/azure/virtual-machines/troubleshooting/troubleshooting-throttling-errors
-    return new ctor(this.auth.credentials, this.auth.subscriptionId, {
-      noRetryPolicy: true, // < -- This removes retry policies so they can be added after the Deserialization
-      requestPolicyFactories: (
-        defaultRequestPolicyFactories: RequestPolicyFactory[],
-      ): RequestPolicyFactory[] => {
-        const policyChain = defaultRequestPolicyFactories;
-
-        if (!this.noRetryPolicy) {
-          // < -- Tests may turn off retry altogether
-          policyChain.push(
-            ...[
-              exponentialRetryPolicy(), // < -- This will not retry a 429 response
-              systemErrorRetryPolicy(),
-              throttlingRetryPolicy(), // < -- This thing will only retry once
-            ],
-          );
-        }
-
-        policyChain.push(bunyanLogPolicy(this.logger));
-
-        return policyChain;
-      },
-    });
-  }
-
-  /**
-   * Iterate all resources of the provided `resourceEndpoint`.
-   *
-   * Sometimes listing fails with a `404` response. In this case, the `callback`
-   * will not be invoked, as if there are no contained resources.
-   *
-   * @param resourceEndpoint a module that supports list()/listNext() or
-   * listAll()/listAllNext()
-   * @param callback a function to receive each resource throughout pagination
-   * @param endpointRatePeriod number of milliseconds over which rate limit
-   * applies
-   */
-  private async iterateAllResources<ServiceClientType, ResourceType>({
-    serviceClient,
-    resourceEndpoint,
-    callback,
-    endpointRatePeriod = 5 * 60 * 1000,
-  }: {
-    serviceClient: ServiceClientType;
-    resourceEndpoint: ListAllResourcesEndpoint | ListResourcesEndpoint;
-    callback: (
-      resource: ResourceType,
-      serviceClient: ServiceClientType,
-    ) => void | Promise<void>;
-    endpointRatePeriod?: number;
-  }): Promise<void> {
-    let nextLink: string | undefined;
-    try {
-      do {
-        const response = await retryResourceRequest(async () => {
-          if ("listAllNext" in resourceEndpoint) {
-            return nextLink
-              ? /* istanbul ignore next: testing iteration might be difficult */
-                await resourceEndpoint.listAllNext<
-                  ResourceListResponse<ResourceType>
-                >(nextLink)
-              : await resourceEndpoint.listAll<
-                  ResourceListResponse<ResourceType>
-                >();
-          } else {
-            return nextLink
-              ? /* istanbul ignore next: testing iteration might be difficult */
-                await resourceEndpoint.listNext<
-                  ResourceListResponse<ResourceType>
-                >(nextLink)
-              : await resourceEndpoint.list<
-                  ResourceListResponse<ResourceType>
-                >();
-          }
-        }, endpointRatePeriod);
-
-        this.logger.info(
-          {
-            resourceCount: response.length,
-            resource: response._response.request.url,
-          },
-          "Received resources for endpoint",
-        );
-
-        for (const e of response) {
-          callback(e, serviceClient);
-        }
-
-        nextLink = response.nextLink;
-      } while (nextLink);
-    } catch (err) {
-      /* istanbul ignore else */
-      if (err.statusCode === 404) {
-        this.logger.warn({ err }, "Resources not found");
-      } else {
-        throw new IntegrationError("Failed to list resources", err, {
-          expose: true,
-        });
-      }
-    }
   }
 }
